+++ app/src/main/java/com/techcourse/Application.java
@@ -10,7 +10,7 @@ public class Application {
 
     private static final Logger log = LoggerFactory.getLogger(Application.class);
 
-    private static final int DEFAULT_PORT = 8080;
+    private static final int DEFAULT_PORT = 8081;
 
     public static void main(final String[] args) throws Exception {
         final int port = defaultPortIfNull(args);
