+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -1,12 +1,7 @@
 package com.techcourse;
 
 import com.interface21.web.WebApplicationInitializer;
-import com.interface21.webmvc.servlet.mvc.AnnotationHandlerAdaptor;
-import com.interface21.webmvc.servlet.mvc.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.HandlerAdaptor;
-import com.interface21.webmvc.servlet.HandlerMapping;
 import jakarta.servlet.ServletContext;
-import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
