+++ app/src/main/java/com/techcourse/ManualHandlerAdaptor.java
@@ -3,8 +3,7 @@ package com.techcourse;
 import com.interface21.webmvc.servlet.HandlerAdaptor;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
+import com.interface21.webmvc.servlet.mvc.Controller;
 import com.interface21.webmvc.servlet.view.type.JspView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
