+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/ForwardController.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.asis;
+package com.interface21.webmvc.servlet.mvc;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
